backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
joblib==1.3.2
python-decouple==3.8
